# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discover:
    get:
      operationId: getDiscoverProfiles
      tags: [discovery]
      summary: Get suggested climbing partners
      parameters:
        - name: gymId
          in: query
          schema:
            type: string
        - name: discipline
          in: query
          schema:
            type: string
            enum: [bouldering, ropes, outdoor, any]
      responses:
        "200":
          description: Suggested profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClimberProfile"
...
  /profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the current climber profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClimberProfile"
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update the current climber profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClimberProfile"
    delete:
      operationId: deleteProfile
      tags: [profile]
      summary: Permanently delete your account and all your data
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /swipes:
    post:
      operationId: createSwipe
      tags: [discovery]
      summary: Like or pass on a climber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeInput"
      responses:
        "200":
          description: Swipe result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResult"
    get:
      operationId: getPendingLikes
      tags: [discovery]
      summary: Get climbers you've liked who haven't liked you back yet
      responses:
        "200":
          description: Pending likes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClimberProfile"
...
  /posts:
    post:
      operationId: createPost
      tags: [feed]
      summary: Create a feed post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /uploads:
    post:
      operationId: uploadImage
      tags: [feed]
      summary: Upload and process an image (HEIC/JPEG/PNG/WebP → JPEG, auto-oriented, resized)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadImagePayload"
      responses:
        "200":
          description: Uploaded image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedImage"
  /profiles/{profileId}/posts:
    get:
      operationId: getProfilePosts
      tags: [feed]
      summary: Get a climber's posts
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
...
  /matches/{matchId}:
    delete:
      operationId: unmatch
      tags: [discovery]
      summary: Unmatch a climber
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unmatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /events:
    get:
      operationId: getEvents
      summary: List nearby climbing events
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [gym, outdoor, trip, all]
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, crew-only, all]
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClimbingEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a climbing event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClimbingEvent"
  /events/{eventId}/rsvp:
    post:
      operationId: rsvpToEvent
      tags: [events]
      summary: RSVP to a climbing event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClimbingEvent"
  /checkins:
    post:
      operationId: createCheckin
      tags: [events]
      summary: Check in at a climbing gym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinInput"
      responses:
        "201":
          description: Check-in created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
  /conversations:
    get:
      operationId: getConversations
      tags: [messaging]
      summary: List direct message conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messaging]
      summary: Start (or get) a conversation with a matched climber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [messaging]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messaging]
      summary: Send a direct message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
...
    ClimberProfile:
      type: object
      required: [id, name, age, location, bio, avatarUrl, disciplines, gyms, availability, matchPercent, openToDating, gear, verified]
      properties:
        id: { type: string }
        name: { type: string }
        age: { type: number }
        location: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        onboarded: { type: boolean }
        disciplines:
          type: array
          items: { type: string }
        gyms:
          type: array
          items: { type: string }
        availability:
          type: array
          items: { type: string }
        matchPercent: { type: number }
        openToDating: { type: boolean }
        gear:
          type: array
          items: { type: string }
        verified: { type: boolean }
        lastActive: { type: string }
        climbingLevel: { type: string }
        avatarPositionX: { type: number }
        avatarPositionY: { type: number }
    ProfileUpdate:
      type: object
      properties:
        name: { type: string }
        age: { type: number }
        location: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        onboarded: { type: boolean }
        disciplines:
          type: array
          items: { type: string }
        gyms:
          type: array
          items: { type: string }
        availability:
          type: array
          items: { type: string }
        openToDating: { type: boolean }
        gear:
          type: array
          items: { type: string }
        climbingLevel: { type: string }
        avatarPositionX: { type: number }
        avatarPositionY: { type: number }
    UploadImagePayload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        kind:
          type: string
          enum: [post, avatar]
    UploadedImage:
      type: object
      required: [url]
      properties:
        url: { type: string }
...
    SwipeInput:
      type: object
      required: [profileId, action]
      properties:
        profileId: { type: string }
        action:
          type: string
          enum: [like, pass]
    SwipeResult:
      type: object
      required: [action, isMatch]
      properties:
        action: { type: string }
        isMatch: { type: boolean }
        match:
          anyOf:
            - $ref: "#/components/schemas/Match"
            - type: "null"
...
    ClimbingEvent:
      type: object
      required: [id, title, type, visibility, dateLabel, timeLabel, location, host, spots, attendees, description, imageUrl, joined]
      properties:
        id: { type: string }
        title: { type: string }
        type:
          type: string
          enum: [gym, outdoor, trip]
        visibility:
          type: string
          enum: [public, crew-only]
        dateLabel: { type: string }
        timeLabel: { type: string }
        location: { type: string }
        host:
          $ref: "#/components/schemas/ClimberProfile"
        spots: { type: number }
        attendees: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        joined: { type: boolean }
    EventInput:
      type: object
      required: [title, type, visibility, dateLabel, timeLabel, location, description, spots]
      properties:
        title: { type: string }
        type:
          type: string
          enum: [gym, outdoor, trip]
        visibility:
          type: string
          enum: [public, crew-only]
        dateLabel: { type: string }
        timeLabel: { type: string }
        location: { type: string }
        description: { type: string }
        spots: { type: number }
        imageUrl: { type: string }
    CheckinInput:
      type: object
      required: [gymId]
      properties:
        gymId: { type: string }
        note: { type: string }
    Checkin:
      type: object
      required: [id, gymId, gymName, createdAt]
      properties:
        id: { type: string }
        gymId: { type: string }
        gymName: { type: string }
        createdAt: { type: string }
        note: { type: string }
    Post:
      type: object
      required: [id, author, postType, createdAt, taggedProfiles]
      properties:
        id: { type: string }
        author:
          $ref: "#/components/schemas/ClimberProfile"
        body:
          type: [string, "null"]
        imageUrl:
          type: [string, "null"]
        postType:
          type: string
          enum: [post, checkin, achievement]
        checkinGymName:
          type: [string, "null"]
        taggedProfiles:
          type: array
          items:
            $ref: "#/components/schemas/ClimberProfile"
        createdAt: { type: string }
    PostInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
        imageUrl:
          type: string
        taggedProfileIds:
          type: array
          items: { type: string }
        postType:
          type: string
          enum: [post, achievement]
...